+++ test/integration/async-job-tracker.test.ts
@@ -31,7 +31,7 @@ function createState() {
 		baseCwd: "/repo",
 		currentSessionId: null,
 		asyncJobs: new Map(),
-		grandTotal: { mainCost: 0, syncCostByRun: new Map(), asyncCostByJob: new Map() },
+		grandTotal: { mainCost: 0, syncCostByRun: new Map(), asyncCostByJob: new Map(), externalCostBySource: new Map() },
 		cleanupTimers: new Map(),
 		lastUiContext: null,
 		poller: null,